driver: Use drm_panel backlight support
There is a new drm_panel_of_backlight() function that handles everything
necessary for external backlight/brightness support introduced in:
  https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=152dbdeab1b2571c107357a7bd59f5cd5be52e30

At the moment it is not intended to be used when the driver itself
manages the backlight (e.g. with MIPI DCS commands). But especially
for the panel drivers generated by lmdpdg it is quite convenient
because there is no need to implement the enable()/disable() functions
when using it.

The documentation for drm_panel->backlight says:

	/**
	 * Backlight device, used to turn on backlight after the call
	 * to enable(), and to turn off backlight before the call to
	 * disable().
	 * backlight is set by drm_panel_of_backlight() and **drivers
	 * shall not assign it.**
	 */

But assigning backlight manually there is working quite well
and allows to simplify the code, so let's ignore it for now.

(I am going to submit a change upstream that allows using it for
 in-driver implementations of backlight as well...)