driver: Mention that lmdpdg was used in the license header
I think it might be fair to mention the device tree that the panel
driver was generated from (IANAL). It also makes it clear that we
might not know much about how the panel works, since the panel
initialization sequence is just generated from the information
available in the device tree.

Everyone is free to remove the notice if they want (I don't care).