Make reset GPIO_ACTIVE_LOW
Many panels have active low reset GPIOs. This can be seen if we keep
reset high after turning the panel on. From a logical perspective this
does not make sense: We should assert reset to actually do the reset,
not to disable it.

Therefore we try check the last element from the reset sequence here.
If it sets the GPIO to 1 (high), we assume that reset is active low.

Also generate this properly in the device tree fragment to avoid
confusion.