driver: Add --backlight-fallback-dcs option
Some panels might be used on multiple devices in different
configurations, using DCS commands or directly with a PWM controller.
In the latter case there is a "backlight = <&...>;" phandle in the DT,
while DCS is handled internally in the panel driver.

Add a --backlight-fallback-dcs option that allows using both cases
dynamically. When a backlight is specified in DT the DCS backlight
is not registered, otherwise the DCS backlight is used.