panel: Chain in `post-panel-on-command`
This is a perfect contended for the `.prepare()` -> `.enable()` split;
on Sony panels the unblank command is typically in `post-panel-on` to
make it happen after the panel is prepared, the bridge is on, and pixel
data is flowing.

Without this pretty much all our panels are generated with a driver that
never unblanks the panel.